trac_alpha_bot/requirements.txt
python-telegram-bot==20.8
websocket-client==1.7.0
